requirements.txt
Flask==3.0.3
pandas==2.0.0
pyngrok==7.1.6
python-dotenv==1.0.1
scikit_learn==1.5.0
